scripts/requirements.txt
tonsdk==1.0.9
httpx==0.19.0
pynacl==1.5.0